commit updates

+++ rl2/agents_v2/preprocessing/tabular.py
@@ -24,7 +24,7 @@ class MABPreprocessing(tc.nn.Module):
         Args:
             prev_action: tc.LongTensor of shape [B, ...]
             prev_reward: tc.FloatTensor of shape [B, ...]
-            prev_done: tc.LongTensor of shape [B, ...]
+            prev_done: tc.FloatTensor of shape [B, ...]
 
         Returns:
             tc.FloatTensor of shape [B, ..., A+2]
@@ -57,7 +57,7 @@ class MDPPreprocessing(tc.nn.Module):
             curr_obs: tc.FloatTensor of shape [B, ..., C, H, W]
             prev_action: tc.LongTensor of shape [B, ...]
             prev_reward: tc.FloatTensor of shape [B, ...]
-            prev_done: tc.LongTensor of shape [B, ...]
+            prev_done: tc.FloatTensor of shape [B, ...]
 
         Returns:
             tc.FloatTensor of shape [B, ..., S+A+2]
